# usagestats: Fix parens in statistics